requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncer==0.0.8
attrs==25.3.0
audioop-lts==0.2.1
autogen==0.8.1
autogen-agentchat==0.4.9.2
autogen-core==0.4.9.2
autogen-ext==0.4.9.2
autogenstudio==0.4.1.11
azure-ai-documentintelligence==1.0.1
azure-ai-inference==1.0.0b9
azure-core==1.32.0
azure-identity==1.21.0
backoff==2.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
cryptography==43.0.3
defusedxml==0.7.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
email_validator==2.2.0
et_xmlfile==2.0.0
fast-depends==2.4.12
fastapi==0.115.11
fastapi-cli==0.0.7
fastapi-sso==0.16.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
gunicorn==22.0.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.63.11
loguru==0.7.3
lxml==5.3.1
Mako==1.3.9
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.0.2
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.32.0
msal-extensions==1.3.1
multidict==6.1.0
numpy==2.2.3
oauthlib==3.2.2
olefile==0.47
openai==1.66.3
openpyxl==3.1.5
opentelemetry-api==1.31.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
pipreqs==0.4.13
pipreqsnb==0.2.4
playwright==1.50.0
propcache==0.3.0
protobuf==5.29.3
psycopg==3.2.6
puremagic==1.28
pyautogen==0.8.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rq==2.1.0
selenium==4.29.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.1
termcolor==2.5.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.0.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.2
yarg==0.1.10
yarl==1.18.3
youtube-transcript-api==1.0.1
zipp==3.21.0
